+++ lib/Model/SmartTransactionsPrepareCallbackUrls.php
@@ -1,0 +1,232 @@
+<?php
+
+namespace Secuconnect\Client\Model;
+
+use \ArrayAccess;
+
+/**
+ * SmartTransactionsPrepareCallbackUrls
+ *
+ * @category    Class
+ * @package     Secuconnect\Client
+ * @author      Swagger Codegen team
+ * @link        https://github.com/swagger-api/swagger-codegen
+ */
+class SmartTransactionsPrepareCallbackUrls implements ArrayAccess
+{
+    const DISCRIMINATOR = null;
+
+    /**
+     * The original name of the model.
+     * @var string
+     */
+    protected static $swaggerModelName = 'SmartTransactionsPrepareCallbackUrls';
+
+    /**
+     * Array of property to type mappings. Used for (de)serialization
+     * @var string[]
+     */
+    protected static $swaggerTypes = [
+        'success_url' => 'string',
+        'failure_url' => 'string'
+    ];
+
+    /**
+     * Array of property to format mappings. Used for (de)serialization
+     * @var string[]
+     */
+    protected static $swaggerFormats = [
+        'success_url' => null,
+        'failure_url' => null
+    ];
+
+    public static function swaggerTypes()
+    {
+        return self::$swaggerTypes;
+    }
+
+    public static function swaggerFormats()
+    {
+        return self::$swaggerFormats;
+    }
+
+    /**
+     * Array of attributes where the key is the local name, and the value is the original name
+     * @var string[]
+     */
+    protected static $attributeMap = [
+        'success_url' => 'success_url',
+        'failure_url' => 'failure_url'
+    ];
+
+    /**
+     * Array of attributes to setter functions (for deserialization of responses)
+     * @var string[]
+     */
+    protected static $setters = [
+        'success_url' => 'setSuccessUrl',
+        'failure_url' => 'setFailureUrl'
+    ];
+
+    /**
+     * Array of attributes to getter functions (for serialization of requests)
+     * @var string[]
+     */
+    protected static $getters = [
+        'success_url' => 'getSuccessUrl',
+        'failure_url' => 'getFailureUrl'
+    ];
+
+    public static function attributeMap()
+    {
+        return self::$attributeMap;
+    }
+
+    public static function setters()
+    {
+        return self::$setters;
+    }
+
+    public static function getters()
+    {
+        return self::$getters;
+    }
+
+    /**
+     * Associative array for storing property values
+     * @var mixed[]
+     */
+    protected $container = [];
+
+    /**
+     * Constructor
+     * @param mixed[] $data Associated array of property values initializing the model
+     */
+    public function __construct(array $data = null)
+    {
+        $this->container['success_url'] = isset($data['success_url']) ? $data['success_url'] : null;
+        $this->container['failure_url'] = isset($data['failure_url']) ? $data['failure_url'] : null;
+    }
+
+    /**
+     * show all the invalid properties with reasons.
+     *
+     * @return array invalid properties with reasons
+     */
+    public function listInvalidProperties()
+    {
+        $invalid_properties = [];
+
+        return $invalid_properties;
+    }
+
+    /**
+     * validate all the properties in the model
+     * return true if all passed
+     *
+     * @return bool True if all properties are valid
+     */
+    public function valid()
+    {
+        return true;
+    }
+
+    /**
+     * Gets success_url
+     * @return string
+     */
+    public function getSuccessUrl()
+    {
+        return $this->container['success_url'];
+    }
+
+    /**
+     * Sets success_url
+     * @param string $success_url Success URL
+     * @return $this
+     */
+    public function setSuccessUrl($success_url)
+    {
+        $this->container['success_url'] = $success_url;
+
+        return $this;
+    }
+
+    /**
+     * Gets failure_url
+     * @return string
+     */
+    public function getFailureUrl()
+    {
+        return $this->container['failure_url'];
+    }
+
+    /**
+     * Sets failure_url
+     * @param string $failure_url Failure URL
+     * @return $this
+     */
+    public function setFailureUrl($failure_url)
+    {
+        $this->container['failure_url'] = $failure_url;
+
+        return $this;
+    }
+    /**
+     * Returns true if offset exists. False otherwise.
+     * @param  integer $offset Offset
+     * @return boolean
+     */
+    public function offsetExists($offset)
+    {
+        return isset($this->container[$offset]);
+    }
+
+    /**
+     * Gets offset.
+     * @param  integer $offset Offset
+     * @return mixed
+     */
+    public function offsetGet($offset)
+    {
+        return isset($this->container[$offset]) ? $this->container[$offset] : null;
+    }
+
+    /**
+     * Sets value based on offset.
+     * @param  integer $offset Offset
+     * @param  mixed   $value  Value to be set
+     * @return void
+     */
+    public function offsetSet($offset, $value)
+    {
+        if (is_null($offset)) {
+            $this->container[] = $value;
+        } else {
+            $this->container[$offset] = $value;
+        }
+    }
+
+    /**
+     * Unsets offset.
+     * @param  integer $offset Offset
+     * @return void
+     */
+    public function offsetUnset($offset)
+    {
+        unset($this->container[$offset]);
+    }
+
+    /**
+     * Gets the string presentation of the object
+     * @return string
+     */
+    public function __toString()
+    {
+        if (defined('JSON_PRETTY_PRINT')) { // use JSON pretty print
+            return json_encode(\Secuconnect\Client\ObjectSerializer::sanitizeForSerialization($this), JSON_PRETTY_PRINT);
+        }
+
+        return json_encode(\Secuconnect\Client\ObjectSerializer::sanitizeForSerialization($this));
+    }
+}